Move definitions module (update imports)
