+++ kissim/definitions.py
